app/requirements_temp.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.7.0
daphne==4.1.2
dj-database-url==2.1.0
Django==5.0.3
django-adminlte3==0.1.6
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
fonttools==4.52.4
gunicorn==23.0.0
html5lib==1.1
idna==3.7
lxml==5.2.2
Markdown==3.6
mysqlclient==2.2.4
oscrypto==1.3.0
pillow==10.2.0
pycparser==2.22
pydyf==0.10.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyMySQL==1.1.0
pypdf==4.2.0
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
openpyxl>=3.1.2 